requirements.txt
django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
whitenoise==6.9.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0